requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
comm==0.1.3
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.14.0
jedi==0.18.2
jupyter-client==8.2.0
jupyter-core==5.3.0
lxml==4.9.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing-extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
zipp==3.15.0
zope.interface==6.0
